Refactored email sending into EmailService and added additional testing
